stripe_payment/requirements.txt
Django==3.2.6
djangorestframework==3.12.4
stripe==2.60